requirements.txt
pandas
beautifulsoup4 
scrapy 
selenium 
requests 
telethon
pyyaml
dbt